General adjustments after sync meeting done.
