+++ Infrastructure/Sources/RocketsClient/Models/RocketDetail+mocks.swift
