+++ Infrastructure/Sources/RocketsClient/Models/RocketDetail.swift
@@ -1,4 +1,3 @@
-import ComposableArchitecture
 import Foundation
 
 public struct RocketDetail: Decodable, Identifiable, Equatable {
